\name{testmat}
\docType{data}
\alias{testmat}
\title{Example results from fastMLA requiring all significance methods}
\description{
This data.frame contains example results from fastMLA. It is comprised of example triplets for all three methods of calculating significance: the full CNM, the simple CNM, and the direct estimate method. It is intended for use with examples using mass.CNM or fastboots.GLA
}
\usage{testmat}
\format{A data frame containing 9 observations of 5 variables (three character("X1 or X2","X2 or X1","X3") and two numeric("est.rho1-rho0"and "GLA.value")}
\source{Generated based on results obtained using the Spellman et al. dataset using methods extended from Ho et al.}
\references{
Yen-Yi Ho, Giovanni Parmigiani, Thomas A Louis, and Leslie M Cope. Modeling liquid association.
Biometrics, 67(1):133-141, 2011.

Paul T Spellman, Gavin Sherlock, Michael Q Zhang, Vishwanath R Iyer, Kirk Anders, Michael B Eisen,
Patrick O Brown, David Botstein, and Bruce Futcher. Comprehensive identification of cell cycle regulated
genes of the yeast saccharomyces cerevisiae by microarray hybridization. Molecular biology of the cell,
9(12):3273-3297, 1998.
}
\keyword{datasets}